requirements.txt
flask
mysql-connector-python-rf
simplejson
requests
MySQL-python
